[Feat] #62 user-info 관련 api 분리. - 경로가 보다 명확해진다고 판단함.
